JM上新Excel导入（空白行判定追加，小数点精度修改）

+++ voyageone-com/src/main/java/com/voyageone/common/util/CommonUtil.java
@@ -237,7 +237,7 @@ public final class CommonUtil {
 
         BigDecimal bd1 = new BigDecimal(doubleValue);
 
-        ret = bd1.setScale(2, BigDecimal.ROUND_HALF_UP).floatValue();
+        ret = bd1.setScale(2, BigDecimal.ROUND_HALF_UP).doubleValue();
 
         return ret;
     }
